requirements.txt
Faker
setuptools
dataclass_csv
jinja2
google-cloud-spanner
